# Fix window update, reduce memory utilization and improve performance