@stuartcarnie 2017-09-12
This also addresses issue #43 

The `Large` benchmark now caps out at 512k