@stuartcarnie 2017-09-12
[This gist][0] shows real-world numbers for Influx Enterprise

**tl;dr**

* total allocations went from 8.02GB down to 2.87MB or a 2,808 times improvement
* in use memory (for large stream) went from 4.01GB down to 815.27KB or a 39,361 times improvement

[0]: https://gist.github.com/stuartcarnie/6ee099d81f03a8f93524c52a8da84fc3