@stuartcarnie 2017-09-13
I will add a upper message type constant, so that the `if` won’t need changing when validating the range and how this issue is typically addressed.