@stuartcarnie 2017-09-13
Here is the original code, where you will notice the call to `newobject`, because the compiler is capturing `s` to make `handler`

```asm
    431            .          . 		// Switch on the type 
    432            .          . 		switch hdr.MsgType() { 
    433            .          . 		case typeData: 
    434          3MB        3MB 			handler = s.handleStreamMessage 
                   .          .  1109973: MOVQ AX, 0(SP)	                                  session.go:434
                 3MB        3MB  1109977: CALL runtime.newobject(SB)	                      session.go:434
                   .          .  110997c: MOVQ 0x8(SP), AX	                                session.go:434
                   .          .  1109981: LEAQ 0x11f98(IP), CX	                            session.go:434
                   .          .  1109988: MOVQ CX, 0(AX)	                                  session.go:434
                   .          .  110998b: TESTB AL, 0(AX)	                                 session.go:434
                   .          .  110998d: MOVL 0x14a4dd(IP), DX	                           session.go:434
                   .          .  1109993: LEAQ 0x8(AX), BX	                                session.go:434
                   .          .  1109997: TESTL DX, DX	                                    session.go:434
                   .          .  1109999: JNE 0x1109a05	                                   session.go:434
                   .          .  110999b: MOVQ 0xb8(SP), BX	                               session.go:434
                   .          .  11099a3: MOVQ BX, 0x8(AX)	                                session.go:434
                   .          .  11099a7: MOVQ 0x50(SP), CX	                               session.go:434
                   .          .  1109a0d: MOVQ BX, 0(SP)	                                  session.go:434
                   .          .  1109a11: MOVQ 0xb8(SP), AX	                               session.go:434
                   .          .  1109a19: MOVQ AX, 0x8(SP)	                                session.go:434
                   .          .  1109a1e: CALL runtime.writebarrierptr(SB)	                session.go:434
                   .          .  1109a23: MOVQ 0x80(SP), AX	                               session.go:434
                   .          .  1109a2b: LEAQ 0x11eee(IP), CX	                            session.go:434
                   .          .  1109a32: MOVQ 0xb8(SP), BX	                               session.go:434
                   .          .  1109a3a: JMP 0x11099a7	                                   session.go:434
                   .          .  1109a3f: LEAQ 0x3d0da(IP), AX	                            session.go:434
```